Move LoadingActions under actions folder
