+++ client/src/actions/SearchActions.js
@@ -1,6 +1,6 @@
 import fetch from 'isomorphic-fetch'
 import _ from 'lodash'
-import { showLoading, hideLoading } from '../loading/LoadingActions'
+import { showLoading, hideLoading } from './LoadingActions'
 import { showErrors } from './ErrorActions'
 import { facetsReceived, clearFacets } from '../search/facet/FacetActions'
 import { assembleSearchRequest } from '../utils/queryUtils'
