+++ client/test/actions/searchActionSpec.js
@@ -1,6 +1,6 @@
 import '../specHelper'
 import * as module from '../../src/actions/SearchActions'
-import { LOADING_SHOW, LOADING_HIDE } from '../../src/loading/LoadingActions'
+import { LOADING_SHOW, LOADING_HIDE } from '../../src/actions/LoadingActions'
 import { FACETS_RECEIVED, CLEAR_FACETS } from '../../src/search/facet/FacetActions'
 import { SET_ERRORS } from '../../src/actions/ErrorActions'
 import configureMockStore from 'redux-mock-store'
